requirements.txt
tqdm==4.36.0
sqlalchemy==1.3.8
azure-storage-blob==2.1.0
pyodbc==4.0.27
rarfile==3.1
py7zr==0.4 
re
csv
subprocess
numpy
pandas
pathlib
zipfile
tarfile
datetime
contextlib

